Upload cloud-storage base project
