+++ jzp-code/boot-demo/src/main/java/com/tutorial/bootDemo/converter/StudentConverter.java
@@ -1,0 +1,22 @@
+package com.tutorial.bootDemo.converter;
+
+import com.tutorial.bootDemo.dao.Student;
+import com.tutorial.bootDemo.dto.StudentDTO;
+
+public class StudentConverter {
+
+    public static StudentDTO convertStudent(Student student) {
+        StudentDTO studentDTO = new StudentDTO();
+        studentDTO.setId(student.getId());
+        studentDTO.setName(student.getName());
+        studentDTO.setEmail(student.getEmail());
+        return studentDTO;
+    }
+
+    public static Student convertStudent(StudentDTO studentDTO) {
+        Student student = new Student();
+        student.setName(studentDTO.getName());
+        student.setEmail(studentDTO.getEmail());
+        return student;
+    }
+}
