+++ jzp-code/boot-demo/src/main/java/com/tutorial/bootDemo/dao/StudentRepository.java
@@ -1,0 +1,12 @@
+package com.tutorial.bootDemo.dao;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface StudentRepository extends JpaRepository<Student, Long> {
+
+    List<Student> findByEmail(String email);
+}
