+++ jzp-code/boot-demo/src/main/java/com/tutorial/bootDemo/dto/StudentDTO.java
@@ -1,0 +1,33 @@
+package com.tutorial.bootDemo.dto;
+
+public class StudentDTO {
+
+    private long id;
+    private String name;
+    private String email;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+}
