--- jzp-code/boot-demo/src/main/java/com/tutorial/boot_demo/BootDemoApplication.java
@@ -1,13 +1,0 @@
-package com.tutorial.boot_demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BootDemoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BootDemoApplication.class, args);
-	}
-
-}
