--- jzp-code/boot-demo/src/main/java/com/tutorial/boot_demo/dao/Student.java
@@ -1,56 +1,0 @@
-package com.tutorial.boot_demo.dao;
-
-import jakarta.persistence.*;
-
-import static jakarta.persistence.GenerationType.IDENTITY;
-
-@Entity
-@Table(name="student")
-public class Student {
-
-    @Id
-    @Column(name="id")
-    @GeneratedValue(strategy = IDENTITY)
-    private long id;
-
-    @Column(name="name")
-    private String name;
-
-    @Column(name="email")
-    private String email;
-
-    @Column(name="age")
-    private int age;
-
-    public String getName() {
-        return name;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
