--- jzp-code/boot-demo/src/main/java/com/tutorial/boot_demo/dao/StudentRepository.java
@@ -1,12 +1,0 @@
-package com.tutorial.boot_demo.dao;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface StudentRepository extends JpaRepository<Student, Long> {
-
-    List<Student> findByEmail(String email);
-}
