--- jzp-code/boot-demo/src/main/java/com/tutorial/boot_demo/dto/StudentDTO.java
@@ -1,33 +1,0 @@
-package com.tutorial.boot_demo.dto;
-
-public class StudentDTO {
-
-    private long id;
-    private String name;
-    private String email;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-}
