--- jzp-code/boot-demo/src/main/java/com/tutorial/boot_demo/service/StudentService.java
@@ -1,15 +1,0 @@
-package com.tutorial.boot_demo.service;
-
-import com.tutorial.boot_demo.dao.Student;
-import com.tutorial.boot_demo.dto.StudentDTO;
-
-public interface StudentService {
-
-    StudentDTO getStudentById(long id);
-
-    Long addNewStudent(StudentDTO studentDTO);
-
-    void deleteStudentById(long id);
-
-    StudentDTO updateStudentById(long id, String name, String email);
-}
