+++ jzp-code/boot-demo/src/test/java/com/tutorial/bootDemo/BootDemoApplicationTests.java
@@ -1,0 +1,13 @@
+package com.tutorial.bootDemo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BootDemoApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
