--- jzp-code/boot-demo/src/test/java/com/tutorial/boot_demo/BootDemoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.tutorial.boot_demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BootDemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
