+++ jzp-code/cloud-storage/cloud-storage-parent/cloud-storage-service/src/main/java/com/xiaomi/service/DataService.java
@@ -1,0 +1,12 @@
+package com.xiaomi.service;
+
+public interface DataService {
+
+    Long insert(String fileName);
+
+    void delete(Long id);
+
+    Long acquire(Long id);
+
+    String rename(Long id);
+}
