+++ jzp-code/cloud-storage/cloud-storage-parent/cloud-storage-service/src/main/java/com/xiaomi/service/FileService.java
@@ -1,0 +1,24 @@
+package com.xiaomi.service;
+
+public class FileService implements DataService{
+
+    @Override
+    public Long insert(String fileName) {
+        return 0L;
+    }
+
+    @Override
+    public void delete(Long id) {
+
+    }
+
+    @Override
+    public Long acquire(Long id) {
+        return 0L;
+    }
+
+    @Override
+    public String rename(Long id) {
+        return "";
+    }
+}
