+++ jzp-code/cloud-storage/cloud-storage-parent/cloud-storage-web/src/main/java/com/xiaomi/controller/TestApplication.java
@@ -1,0 +1,11 @@
+package com.xiaomi.controller;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(TestApplication.class, args);
+    }
+}
