+++ jzp-code/cloud-storage/cloud-storage-parent/cloud-storage-web/src/main/java/com/xiaomi/controller/TestController.java
@@ -1,0 +1,13 @@
+package com.xiaomi.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class TestController {
+
+    @GetMapping("/hello")
+    public String hello(){
+        return "Hello, xiaomi";
+    }
+}
